implmented expo router navigation
