--- app/HomePage.jsx
@@ -1,38 +1,0 @@
-import AntDesign from "@expo/vector-icons/AntDesign";
-import React from "react";
-import { StyleSheet, TouchableOpacity, View } from "react-native";
-import RecipeList from "./RecipeList";
-import CameraView from "./CameraView"
-
-function HomePage() {
-  return (
-    // <View style={styles.homePageContainer}>
-    //   <RecipeList />
-    //   <TouchableOpacity style={styles.buttonContainer}>
-    //     <AntDesign
-    //       name="pluscircle"
-    //       size={60}
-    //       color="black"
-    //       style={styles.button}
-    //     />
-    //   </TouchableOpacity>
-    // </View>
-
-    <CameraView />
-  );
-}
-
-const styles = StyleSheet.create({
-  homePageContainer: {
-    flex: 1,
-    flexDirection: "column",
-    padding: 15,
-    backgroundColor: "white",
-  },
-  buttonContainer: {
-    marginVertical: "auto",
-    alignItems: "center",
-  },
-});
-
-export default HomePage;
