+++ app/_layout.jsx
@@ -1,0 +1,8 @@
+import { Stack } from "expo-router";
+
+export default function RootLayout() {
+  return <Stack>
+    <Stack.Screen name="index" />
+    <Stack.Screen name="camera" />
+  </Stack>;
+}
