+++ app/camera.jsx
@@ -1,4 +1,4 @@
-import Camera from "./Camera";
+import Camera from "../components/Camera";
 
 function CameraView() {
   return <Camera />;
