--- app/index.js
@@ -1,8 +1,0 @@
-import React from "react";
-import HomePage from "./HomePage";
-
-function App() {
-  return <HomePage />;
-}
-
-export default App;
