+++ package.json
@@ -17,12 +17,12 @@
     "@react-navigation/native": "^7.1.6",
     "expo": "~53.0.16",
     "expo-blur": "~14.1.5",
-    "expo-constants": "~17.1.6",
+    "expo-constants": "~17.1.7",
     "expo-font": "~13.3.2",
     "expo-haptics": "~14.1.4",
     "expo-image": "~2.3.2",
-    "expo-linking": "~7.1.6",
-    "expo-router": "~5.1.2",
+    "expo-linking": "~7.1.7",
+    "expo-router": "~5.1.3",
     "expo-splash-screen": "~0.30.9",
     "expo-status-bar": "~2.2.3",
     "expo-symbols": "~0.4.5",
