+++ tsconfig.json
@@ -13,5 +13,5 @@
     "**/*.tsx",
     ".expo/types/**/*.ts",
     "expo-env.d.ts"
-, "app/index.js"  ]
+, "app/index.js", "app/_layout.jsx"  ]
 }
